% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lineqBAGP.R
\name{augment.lineqBAGP}
\alias{augment.lineqBAGP}
\title{Augmenting Method for the \code{"lineqBAGP"} S3 Class}
\usage{
\method{augment}{lineqBAGP}(x, ...)
}
\arguments{
\item{x}{An object with class \code{lineqBAGP}.}

\item{...}{further arguments passed to or from other methods.}
}
\value{
An expanded \code{"lineqBAGP"} An object with the following additional elements.

\item{Phi}{a matrix corresponding to the hat basis functions.
The basis functions are indexed by rows}
\item{kernel$Param}{an object indicating for each variable the associated kernel} #will evolve
\item{Gamma.var}{a list of list of matrix where Gamma.var[[j]][[k]][i] is 
the covariance matrix of the 1D GP with kernel kernel$Param[[j]][[k]] for the points 
\eqn{x^{(1)}_{\sigma(i)},..., x^{n}_{\sigma(i)}}}.
\item{(Lambda,lb,ub)}{the linear system of inequalities.}
\item{...}{further parameters passed to or from other methods.}
}
\description{
Augmenting method for the \code{"lineqBAGP"} S3 class.
}
\details{
Some paramaters of the finite-dimensional GP with linear inequality
constraints are computed. Here, \eqn{\boldsymbol{\xi}}{\xi} is a centred Gaussian
vector with covariance \eqn{\boldsymbol{\Gamma}}{\Gamma}, s.t.
\eqn{\boldsymbol{\Phi} \boldsymbol{\xi} = \boldsymbol{y}}{\Phi \xi = y}
(interpolation constraints) and
\eqn{\boldsymbol{l} \leq \boldsymbol{\Lambda} \boldsymbol{\xi} \leq \boldsymbol{u}}{lb \le \Lambda \xi \le ub}
(inequality constraints).
}
\examples{

# synthetic data
targetFun <- function(x)  return(x[, 1]*x[, 3] + x[, 2])

d <- 3 # nb of dimensions
nblocks <- 2 # nb of blocks

nbdesign <- 4
xdesign <- matrix(runif(nbdesign*d), nbdesign, d)
ydesign <- targetFun(xdesign)

# creating the model
model <- create(class = "lineqBAGP", x = xdesign, y = ydesign,
                constrType = rep("monotonicity", nblocks),
                partition = list(c(1, 3), 2),
                subdivision_size = list(c(3, 5), 2))
str(model)
                
# computing some matrices used in the "lineqBAGP" model
model <- augment(model)
str(model)

}
\references{
A. F. Lopez-Lopera (2019),
"Gaussian process modelling under inequality constraints".
\emph{PhD thesis, Mines Saint-Etienne}
<https://tel.archives-ouvertes.fr/tel-02863891>
}
\seealso{
\code{\link{create.lineqBAGP}}, \code{\link{predict.lineqBAGP}},
         \code{\link{simulate.lineqBAGP}}
}
\author{
M. Deronzier and A. F. Lopez-Lopera
}
