% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lineqMaxModBAGP.R
\name{BAGPMaxMod}
\alias{BAGPMaxMod}
\title{MaxMod algorithm for \code{"lineqGP"} and \code{"lineqBAGP"} Models}
\usage{
BAGPMaxMod(
  model,
  xtest = 0,
  max_iter = 10 * ncol(model$x),
  reward_new_knot,
  reward_new_dim = 1e-09,
  print_iter = FALSE,
  nClusters = 1,
  save_history = FALSE,
  constrType
)
}
\arguments{
\item{model}{an object with class \code{lineqGP} or \code{lineqBAGP}}

\item{xtest}{test data for assessing the modification of the MAP estimate}

\item{max_iter}{an integer corresponding to number of iterations}

\item{reward_new_knot}{a number corresponding to the reward of adding a new knot in an existing dimension}

\item{reward_new_dim}{a number corresponding to the reward of adding a new dimension}

\item{print_iter}{a logical variable to print results at each iteration}

\item{nClusters}{an integer corresponding to the number of clusters}

\item{save_history}{a logical variable to save the model at each iteration}

\item{tol}{a number corresponding to the tolerance of algorithm.
The algorithm stops if the (MaxMod criterion)/norm(pred_old) < tol}
}
\value{
an object with class \code{lineqGP} or \code{lineqBAGP} containing the resulting model
}
\description{
A wrapper function for the maximum Modification algorithm.
}
\author{
A. F. Lopez-Lopera
}
