% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilityFunctions.R
\name{Bijection}
\alias{Bijection}
\title{Useful functions for \code{"lineqGP"} Models.}
\usage{
Bijection(partition, i)
}
\arguments{
\item{sub_partition}{:list[seq] A sub-partition for first input and locations}

\item{subdivision}{:list[list[seq]] The subdivisions for each axe subdivision[[j]][[k]]}

\item{subdivision_taille}{list[seq] subdivision_size[[j]][k] =length(subdivision[[j]][[k]])}

\item{dim_block}{:seq, the size of the blocks, dim_block[j]=length(partition[[j]])
will give the subdivision of $[0,1]$ for the variable partition[[j]][[k]]}

\item{X}{matrix such that X[,i] is an observation $x^{(i)}$}

\item{x}{float supposed to be a real number in $[0,1]$}

\item{Y}{the vector of observation Y[i]=f(X[,i])}

\item{A, B}{list["vector"] Matrices per block of type}

\item{d}{a number corresponding to the dimension of the input space.}
}
\value{
Kernel matrix \eqn{K(x_1,x_2)}{K(x1,x2)}
(or \eqn{K(x_1,x_1)}{K(x1,x1)} if \eqn{x_2}{x2} is not defined).

Bijection will give the other one
}
\description{
Manipulation of blocks matrices for fast computation \code{"lineqGP"} models.
}
\examples{
x <- seq(0, 1, 0.01)
K <- kernCompute(x, type = "gaussian", par =  c(1, 0.1))
image(K, main = "covariance matrix")

}
\seealso{
\code{\link{k1gaussian}}, \code{\link{k1matern52}}, \code{\link{k1matern32}}, \code{\link{k1exponential}}
}
\author{
A. F. Lopez-Lopera, M. Deronzier
}
