% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lineqBAGPutils.R
\name{block_chol}
\alias{block_chol}
\title{Elementwise Cholesky of a list of matrices}
\usage{
block_chol(A)
}
\arguments{
\item{A}{A list containing inversible matrices.}
}
\value{
A list containing the elementwise Cholesky decomposition of \eqn{A}.
}
\description{
Compute the elementwise Cholesky of a list of matrices.
}
\author{
M. Deronzier and A. F. Lopez-Lopera
}
