% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lineqBAGPutils.R
\name{block_chol2inv}
\alias{block_chol2inv}
\title{Elementwise inversion of a list of Cholesky matrices}
\usage{
block_chol2inv(A)
}
\arguments{
\item{A}{A list containing Cholesky matrices.}
}
\value{
A list containing the elementwise inverse of \eqn{A}.
}
\description{
Compute the elementwise inversion a list of Cholesky matrices via 
Cholesky decomposition.
}
\author{
M. Deronzier and A. F. Lopez-Lopera
}
