% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lineqBAGPutils.R
\name{block_compute}
\alias{block_compute}
\title{Computation of elementwise operations of lists of matrices}
\usage{
block_compute(
  A,
  operation = c("sum", "prod", "transpose", "chol", "inv", "chol2inv", "scalarMatMul",
    "det"),
  B = NULL,
  alpha = 1,
  beta = 1
)
}
\arguments{
\item{A}{A list containing matrices.}

\item{operation}{A character string with the operation to be applied.
Options: "sum", "prod", "transpose", "chol", "inv", "chol2inv", "scalarMatMul".}

\item{B}{A list containing matrices.}

\item{alpha}{A real number.}

\item{beta}{A real number.}
}
\value{
A list after applying the elementwise operation.
}
\description{
Compute an elementwise operation to a list (or two lists) of matrices.
}
\author{
M. Deronzier and A. F. Lopez-Lopera
}
