% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lineqBAGPutils.R
\name{block_inv}
\alias{block_inv}
\title{Elementwise inversion of a list of matrices}
\usage{
block_inv(A)
}
\arguments{
\item{A}{A list containing inversible matrices.}
}
\value{
A list containing the elementwise inverse of \eqn{A}.
}
\description{
Compute the elementwise inversion of a list of matrices via Cholesky decomposition.
}
\author{
M. Deronzier and A. F. Lopez-Lopera
}
