% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilityFunctions.R
\name{block_inv_lemma}
\alias{block_inv_lemma}
\title{The type in this function is to deal with a particular case where $Z=\tau^2 I_n$
To make computations faster
Block case, be careful about matrices we use}
\usage{
block_inv_lemma(W, inv_Z, U, V = NULL)
}
\description{
The type in this function is to deal with a particular case where $Z=\tau^2 I_n$
To make computations faster
Block case, be careful about matrices we use
}
