% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lineqBAGPutils.R
\name{block_mul}
\alias{block_mul}
\title{Elementwise scalar-matrix product of a list of matrices}
\usage{
block_mul(alpha, A)
}
\arguments{
\item{alpha}{A real number.}

\item{A}{A list containing matrices.}
}
\value{
A list containing the elementwise scalar-matrix product
\eqn{\alpha A_j \ \forall j = 1, \ldots, n}.
}
\description{
Compute the elementwise scalar-matrix product of a list of matrices.
}
\author{
M. Deronzier and A. F. Lopez-Lopera
}
