% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lineqBAGPutils.R
\name{Block_Phi}
\alias{Block_Phi}
\title{Hat basis functions per variable}
\usage{
Block_Phi(subdivision, partition, x)
}
\arguments{
\item{subdivision}{A list containing the subdivisions for each variable:
\code{subdivision[[j]][[k]]} denotes the subdivision of the block \eqn{j} and
local input variable \eqn{k}.}

\item{partition}{A list containing the partition of set \eqn{\{1, \cdots, d\}}.}

\item{x}{An \eqn{n \times d} matrix with the input data.

\code{subdivision[[j]][[k]]} denotes the subdivision of the block \eqn{j} and
local input variable \eqn{k}.}
}
\value{
A list containing the hat basis functions for each block.
}
\description{
Compute the hat basis functions for each input variable.
}
\author{
M. Deronzier and A. F. Lopez-Lopera
}
