% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilityFunctions.R
\name{block_prod}
\alias{block_prod}
\title{Here are some basic operations on the block matrices}
\usage{
block_prod(A, B)
}
\value{
block_operation input: matrix represented per blocks output:operation on matrix
}
\description{
Here are some basic operations on the block matrices
}
