% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lineqBAGPutils.R
\name{block_scalar}
\alias{block_scalar}
\title{Name}
\usage{
block_scalar(A, B)
}
\arguments{
\item{A}{A list containing matrices}

\item{B}{A list containing matrices}
}
\value{
A matrix object
}
\description{
description
}
\author{
M. Deronzier and A. F. Lopez-Lopera
}
