% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lineqBAGPutils.R
\name{block_sum}
\alias{block_sum}
\title{Elementwise linear combination of two lists of matrices}
\usage{
block_sum(A, B, alpha = 1, beta = 1)
}
\arguments{
\item{A}{A list containing matrices.}

\item{B}{A list containing matrices.}

\item{alpha}{A real number.}

\item{beta}{A real number.}
}
\value{
A list containing the elementwise linear combination
\code{alpha x A[[j]] + beta x B[[j]]}.
}
\description{
Compute the elementwise linear combination of two lists of matrices.
}
\author{
M. Deronzier and A. F. Lopez-Lopera
}
