% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lineqBAGPutils.R
\name{block_to_matrix}
\alias{block_to_matrix}
\title{Block operation wrapper function}
\usage{
block_to_matrix(A, type = c("bdiag", "cbind", "rbind", "sum"))
}
\arguments{
\item{A}{A list containing matrices.}

\item{type}{A string describing the type of operation to be applied for all the blocks.}
}
\value{
A matrix.
}
\description{
The additivity gives us block matrices, hence we can accelerate
the computations this is the role of algorithm in this section to facilitate
the manipulation of matrices of the type \code{list["vector"]} 
\eqn{\{1, \cdots, D \}} induces a natural bijection from
\eqn{\{ (j,k), j \in \{1, \cdots, J \}, k \in \{1,\cdots, \mathcal{J}_b\} \} }
into \eqn{\{1, \cdots, D\}} there is an straight one, \code{partition[[j]][[k]]}
gives the element \eqn{i} and ...
}
\author{
M. Deronzier and A. F. Lopez-Lopera
}
