% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilityFunctions.R
\name{block_to_vect}
\alias{block_to_vect}
\title{The additivity gives us block matrices, hence we can accelerate the computations this is the role of
algorithm in this section to facilitate the manipulation of matrices of the type list["vector"]
The following functions are meant to make matricial computations with matrices per blocks.}
\usage{
block_to_vect(A, type = "diagonal")
}
\value{
block_to_vect input: matrix represented per blocks output: matrix
}
\description{
The additivity gives us block matrices, hence we can accelerate the computations this is the role of
algorithm in this section to facilitate the manipulation of matrices of the type list["vector"]
The following functions are meant to make matricial computations with matrices per blocks.
}
