% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lineqBAGPutils.R
\name{block_transpose}
\alias{block_transpose}
\title{Elementwise transpose of a list of matrices}
\usage{
block_transpose(A)
}
\arguments{
\item{A}{A list containing matrices.}
}
\value{
A list containing the elementwise transpose of \eqn{A}.
}
\description{
Compute the elementwise transpose of a list of matrices.
}
\author{
M. Deronzier and A. F. Lopez-Lopera
}
