% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lineqBAGPutils.R
\name{changement_basis_matrix}
\alias{changement_basis_matrix}
\title{Creation of changing basis matrices in the general case}
\usage{
changement_basis_matrix(partition1, partition2, subdivision1, subdivision2)
}
\arguments{
\item{partition1, }{partition2 two partitions}

\item{subdivision1, subdivision2}{two subdivisions adapted to the partitions we need}
}
\value{
P a changment basis matrix from the basis generated by (partition1, subdivision1) to 
the basis generated by (partition2, subdivision2)
}
\description{
The algorithm explained in the proof of the article.
}
\author{
M. Deronzier and A. F. Lopez-Lopera
}
