% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lineqBAGPutils.R
\name{changment_basis_inclusion}
\alias{changment_basis_inclusion}
\title{Creation of changing basis matrices in case 2 of the article}
\usage{
changment_basis_inclusion(partition1, partition2, subdivision2)
}
\arguments{
\item{partition1, partition2}{two partitions}

\item{subdivision1, subdivision2}{two subdivisions adapted to the partitions we need}
}
\value{
Boolean indicating if we can make a changment basis
}
\description{
This algorithm make the changment basis corresponding to the case 2
of the article.
}
\author{
M. Deronzier and A. F. Lopez-Lopera
}
