% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lineqBAGPutils.R
\name{changment_basis_knots}
\alias{changment_basis_knots}
\title{Creation of changing basis matrices in case 1 of the article}
\usage{
changment_basis_knots(subdivision1, subdivision2)
}
\arguments{
\item{subdivision1, subdivision2}{two subdivisions adapted to the partitions we need}

\item{partition1, partition2}{two partitions}
}
\value{
Boolean indicating if we can make a changment basis
}
\description{
This algorithm make the changement basis corresponding of the case 1
of the article.
}
\author{
M. Deronzier and A. F. Lopez-Lopera
}
