% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lineqBAGPutils.R
\name{condition_changment_basis}
\alias{condition_changment_basis}
\title{Algorithm that check if you can make a changment basis.}
\usage{
condition_changment_basis(partition1, partition2, subdivision1, subdivision2)
}
\arguments{
\item{partition1, partition2}{two partitions}

\item{subdivision1, subdivision2}{two subdivisions adapted to the partitions we need}
}
\value{
boolean indicating if we can make a changment basis
}
\description{
This algorithm check if we are in the hypotheses of the proposition
to make a changment basis.
}
\author{
M. Deronzier and A. F. Lopez-Lopera
}
