% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lineqMaxModBAGP.R
\name{construct_t}
\alias{construct_t}
\title{Choice on the posittion of the node}
\usage{
construct_t(t, model, choice, reward_new_knot = 1e-06, Nscale = 1)
}
\arguments{
\item{t}{the position of the refinement}

\item{model}{an object with clall \code{lineqBAGP}}

\item{choice}{the choice made for the variable where there will be a refinement of the subdivision}

\item{reward_new_knot}{a number corresponding to the reward of adding a new knot in an existing dimension}

\item{NscaLe}{the number of knots we want to add}
}
\value{
the modification of the MAP estimate after adding a new knot
}
\description{
Compute the modification of the MAP estimate according to the MaxMod criterion
}
\references{
F. Bachoc, A. F. Lopez-Lopera, and O. Roustant (2020),
"Sequential construction and dimension reduction of Gaussian processes under inequality constraints".
\emph{ArXiv e-prints}
<arXiv:2009.04188>
}
\author{
A. F. Lopez-Lopera
}
