% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lineqBAGP.R
\name{create.lineqBAGP}
\alias{create.lineqBAGP}
\title{Creation Method for the \code{"lineqBAGP"} S3 Class}
\usage{
\method{create}{lineqBAGP}(x, y, constrType, partition = as.list(seq(ncol(x))), subdivision)
}
\arguments{
\item{x}{A vector or matrix with the input data. The dimensions should be indexed by columns}

\item{y}{A vector with the output data}

\item{constrType}{A character string corresponding to the type of the inequality constraint}

\item{partition}{A list containing the partion of set \eqn{\{1, \cdots, D\}}}

\item{m}{A scalar or vector corresponding to the number of knots per dimension.}
}
\value{
A list with the following elements.
\item{x,y,constrType}{see \bold{Arguments}}
\item{d}{a number corresponding to the input dimension}
\item{constrIdx}{for d > 1, a integer vector with the indices of active constrained dimensions}
\item{constrParam}{constraint inequalities for each dimension}
\item{varnoise}{a scalar with noise variance}
\item{subdivision}{a list of list containing sequence of subdivision of segment $[0,1]$}
\item{localParam}{a list with specific parameters required for \code{"lineqBAGP"} models:
\code{m} (number of basis functions), \code{sampler}, and \code{samplingParams}.
See \code{\link{simulate.lineqBAGP}}}
\item{kernParam}{a list with the kernel parameters: \code{par} (kernel parameters), \code{type}, \code{nugget}.
See \code{\link{kernCompute}}}
\item{bounds}{the limit values if \code{constrType = "boundedness"}.}
\item{(Lambda,lb,ub)}{the linear system of inequalities if \code{constrType = "linear"}}
}
\description{
Creation method for the \code{"lineqBAGP"} S3 class.
}
\examples{

# synthetic data
targetFun <- function(x)  return(x[, 1]*x[, 3] + x[, 2])

d <- 3 # nb of dimensions
nblocks <- 2 # nb of blocks

nbdesign <- 4
xdesign <- matrix(runif(nbdesign*d), nbdesign, d)
ydesign <- targetFun(xdesign)

# creating the model
model <- create(class = "lineqBAGP", x = xdesign, y = ydesign,
                constrType = rep("monotonicity", nblocks),
                partition = list(c(1,3), 2),
                subdivision = list(list(c(0,0.5,1), c(0,0.25,0.5,0.75,0.1)), list(c(0,1)))
str(model)

}
\references{
A. F. Lopez-Lopera (2019),
"Gaussian process modelling under inequality constraints".
\emph{PhD thesis, Mines Saint-Etienne}
<https://tel.archives-ouvertes.fr/tel-02863891>
}
\seealso{
\code{\link{augment.lineqBAGP}}, \code{\link{predict.lineqBAGP}}, \code{\link{simulate.lineqBAGP}}
}
\author{
M. Deronzier and A. F. Lopez-Lopera
}
