% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lineqGPutils.R
\name{create}
\alias{create}
\title{Model Creations}
\usage{
create(class, ...)
}
\arguments{
\item{class}{a character string corresponding to the desired class.}

\item{...}{further arguments passed to or from other methods.
(see, e.g., \code{\link{create.lineqGP}})}
}
\value{
A model object created according to its \code{class}.
}
\description{
Wrapper function for creations of model functions.
The function invokes particular methods which depend on the \code{class}
of the first argument.
}
\examples{
## Not run:
model <- list()
model2 <- create(class = "ClassName", model)
model2
## End(Not run)

}
\seealso{
\code{\link{augment}}, \code{\link{predict}},
         \code{\link{simulate}}
}
\author{
A. F. Lopez-Lopera
}
