% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilityFunctions.R
\name{create_conditional_mean_2}
\alias{create_conditional_mean_2}
\title{Fast computation when m<n and matrix is identity}
\usage{
create_conditional_mean_2(K, Phi, Y, tau)
}
\description{
Fast computation when m<n and matrix is identity
}
