% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lineqMaxModAGP.R, R/lineqMaxModBAGP.R
\name{distAdditiveModes}
\alias{distAdditiveModes}
\title{Modification of the MAP estimate (\code{"lineqAGP"})}
\usage{
distAdditiveModes(u, model, pred, xtest, idx_add, reward_new_knot = 1e-06)

distAdditiveModes(u, model, pred, xtest, idx_add, reward_new_knot = 1e-06)
}
\arguments{
\item{u}{an extended sequence corresponding to the values of the knots}

\item{model}{an object with class \code{lineqAGP}}

\item{pred}{an predictive model with class \code{lineqAGP}}

\item{xtest}{test input data}

\item{idx_add}{index of the dimension that will be activated}

\item{reward_new_knot}{a number corresponding to the reward of adding a new knot in an existing dimension}
}
\value{
the modification of the MAP estimate after adding a new knot

the modification of the MAP estimate after adding a new knot
}
\description{
Compute the modification of the MAP estimate according to the MaxMod criterion

Compute the modification of the MAP estimate according to the MaxMod criterion
}
\references{
F. Bachoc, A. F. Lopez-Lopera, and O. Roustant (2020),
"Sequential construction and dimension reduction of Gaussian processes under inequality constraints".
\emph{ArXiv e-prints}
<arXiv:2009.04188>

F. Bachoc, A. F. Lopez-Lopera, and O. Roustant (2020),
"Sequential construction and dimension reduction of Gaussian processes under inequality constraints".
\emph{ArXiv e-prints}
<arXiv:2009.04188>
}
\author{
A. F. Lopez-Lopera
}
