% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lineqMaxModGP.R
\name{distModes}
\alias{distModes}
\title{Modification of the MAP estimate (\code{"lineqGP"})}
\usage{
distModes(u, model, pred, xtest, idx_add, reward_new_knot = 1e-06, Nscale = 1)
}
\arguments{
\item{u}{an extended sequence corresponding to the values of the knots}

\item{model}{an object with class \code{lineqGP}}

\item{pred}{an predictive model with class \code{lineqGP}}

\item{xtest}{test input data}

\item{idx_add}{index of the dimension that will be activated}

\item{reward_new_knot}{a number corresponding to the reward of adding a new knot in an existing dimension}

\item{Nscale}{an integer corresponding to the number of new added knots}
}
\value{
the modification of the MAP estimate after adding a new knot
}
\description{
Compute the modification of the MAP estimate according to the MaxMod criterion
}
\references{
F. Bachoc, A. F. Lopez-Lopera, and O. Roustant (2020),
"Sequential construction and dimension reduction of Gaussian processes under inequality constraints".
\emph{ArXiv e-prints}
<arXiv:2009.04188>
}
\author{
A. F. Lopez-Lopera
}
