% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lineqGPutils.R
\name{errorMeasureRegress}
\alias{errorMeasureRegress}
\title{Error Measures for GP Models.}
\usage{
errorMeasureRegress(
  y,
  ytest,
  mu,
  varsigma,
  type = "all",
  control = list(nsigma = 1.96)
)
}
\arguments{
\item{y}{a vector with the output observations used for training.}

\item{ytest}{a vector with the output observations used for testing.}

\item{mu}{a vector with the posterior mean.}

\item{varsigma}{a vector with the posterior variances.}

\item{type}{a character string corresponding to the type of the measure.}

\item{control}{an optional list with parameters to be passed (e.g. cia: "nsigma").}
}
\value{
The values of the error measures.
}
\description{
Compute error measures for GP models:
mean absulte error (\code{"mae"}), mean squared error (\code{"mse"}),
standardised mse (\code{"smse"}), mean standardised log loss (\code{"msll"}),
Q2 (\code{"q2"}), predictive variance adequation (\code{"pva"}),
confidence interval accuracy (\code{"cia"}).
}
\examples{
# generating the toy example
n <- 100
w <- 4*pi
x <- seq(0, 1, length = n)
y <- sin(w*x)

# results with high-level noises generating the toy example
nbsamples <- 100
set.seed(1)
ynoise <- y + matrix(rnorm(n*nbsamples, 0, 10), ncol = nbsamples)
mu <- apply(ynoise, 1, mean)
sigma <- apply(ynoise, 1, sd)
matplot(x, ynoise, type = "l", col = "gray70")
lines(x, y, lty = 2, col = "red")
lines(x, mu, col = "blue")
lines(x, mu+1.98*sigma, lty = 2)
lines(x, mu-1.98*sigma, lty = 2)
legend("topright", c("target", "mean", "confidence", "samples"),
       lty = c(2,1,2,1), col = c("red", "blue", "black", "gray70"))
t(errorMeasureRegress(y, y, mu, sigma^2))

# results with low-level noises generating the toy example
set.seed(1)
ynoise <- y + matrix(rnorm(n*nbsamples, 0, 0.05), ncol = nbsamples)
mu <- apply(ynoise, 1, mean)
sigma <- apply(ynoise, 1, sd)
matplot(x, ynoise, type = "l", col = "gray70")
lines(x, y, lty = 2, col = "red")
lines(x, mu, col = "blue")
lines(x, mu+1.98*sigma, lty = 2)
lines(x, mu-1.98*sigma, lty = 2)
legend("topright", c("target", "mean", "confidence", "samples"),
       lty = c(2,1,2,1), col = c("red", "blue", "black", "gray70"))
t(errorMeasureRegress(y, y, mu, sigma^2))

}
\references{
C. E. Rasmussen, and C. K. I. Williams (2005),
"Gaussian Processes for Machine Learning (Adaptive Computation and Machine Learning)".
\emph{The MIT Press}.

F. Bachoc (2013),
"Cross validation and maximum likelihood estimations of hyper-parameters of
Gaussian processes with model misspecification".
\emph{Computational Statistics & Data Analysis},
66:55-69.
}
\seealso{
\code{\link{errorMeasureRegressMC}}
}
\author{
A. F. Lopez-Lopera
}
