% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lineqBAGPutils.R
\name{Gamma_var}
\alias{Gamma_var}
\title{Covariance matrix per variable}
\usage{
Gamma_var(subdivision, par, type)
}
\arguments{
\item{subdivision}{A list containing the subdivisions for each variable:
\code{subdivision[[j]][[k]]} denotes the subdivision of the block \eqn{j} and
local input variable \eqn{k}.}

\item{par}{A list containing the values of the kernel parameters (variance, lengthscales).
\code{par[[j]][l]} denotes the \eqn{l}-th parameter of the block \eqn{j}.}

\item{type}{A character string corresponding to the type of the kernel.
Options: "gaussian", "matern32", "matern52", "exponential".}
}
\value{
A list containing the covariance matrix for each variable. The matrices
are indexed per blocks.
}
\description{
Compute the covariance matrix for each input variable.
}
\author{
M. Deronzier and A. F. Lopez-Lopera
}
