% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lineqBAGPutils.R
\name{Gamma_var_to_tensor}
\alias{Gamma_var_to_tensor}
\title{Covariance matrix per block}
\usage{
Gamma_var_to_tensor(Gamma.perVar)
}
\arguments{
\item{Gamma.perVar}{A list containing the covariance matrix for each variable.
The matrices are indexed per blocks.}
}
\value{
A list containing the covariance matrix for each block.
}
\description{
Compute the covariance matrix for each block.
}
\author{
M. Deronzier and A. F. Lopez-Lopera
}
