% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lineqGPutils.R
\name{ggplot.lineqDGP}
\alias{ggplot.lineqDGP}
\title{GGPlot for the \code{"lineqDGP"} S3 Class}
\usage{
\method{ggplot}{lineqDGP}(data, mapping, ...)
}
\arguments{
\item{data}{an object with \code{lineqDGP} S3 class.}

\item{mapping}{not used.}

\item{...}{further arguments passed to or from other methods.}
}
\value{
GGPlot with the \code{"lineqDGP"} model.
}
\description{
GGPlot for the \code{"lineqDGP"} S3 class.
See \code{\link{ggplot.lineqGP}} for more details.
}
\seealso{
\code{\link{ggplot.lineqGP}}, \code{\link{ggplot}}
}
\author{
A. F. Lopez-Lopera.
}
