% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lineqBAGPutils.R
\name{Gram}
\alias{Gram}
\title{Gram matrix for the 1 dimensional basis functions (\code{"lineqBAGP"})}
\usage{
Gram(subdivision, J = length(subdivision))
}
\arguments{
\item{subdivision}{the multiples subdivisions on the variable}

\item{J}{number of blocks}
}
\value{
Gram matrix of the basis functions
}
\description{
Compute the Gram matrix of the basis functions for \code{"lineqBAGP"} models
}
\references{
F. Bachoc, A. F. Lopez-Lopera, and O. Roustant (2020),
"Sequential construction and dimension reduction of Gaussian processes under inequality constraints".
\emph{ArXiv e-prints}
<arXiv:2009.04188>
}
\author{
M. Deronzier and A. F. Lopez-Lopera
}
