% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lineqMaxModGP.R
\name{GramMatrixPhi}
\alias{GramMatrixPhi}
\title{Gram matrix of the basis functions (\code{"lineqGP"})}
\usage{
GramMatrixPhi(u)
}
\arguments{
\item{u}{a sequence corresponding to the values of the knots}
}
\value{
Gram matrix of the basis functions
}
\description{
Compute the Gram matrix of the basis functions for \code{"lineqGP"} models
}
\references{
F. Bachoc, A. F. Lopez-Lopera, and O. Roustant (2020),
"Sequential construction and dimension reduction of Gaussian processes under inequality constraints".
\emph{ArXiv e-prints}
<arXiv:2009.04188>
}
\author{
A. F. Lopez-Lopera
}
