% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lineqMaxModAGP.R, R/lineqMaxModBAGP.R
\name{GramMatrixPhiAdditive}
\alias{GramMatrixPhiAdditive}
\title{Gram matrix of the basis functions (\code{"lineqAGP"})}
\usage{
GramMatrixPhiAdditive(u)

GramMatrixPhiAdditive(u)
}
\arguments{
\item{u}{a sequence corresponding to the values of the knots}
}
\value{
Gram matrix of the basis functions

Gram matrix of the basis functions
}
\description{
Compute the Gram matrix of the basis functions for \code{"lineqAGP"} models

Compute the Gram matrix of the basis functions for \code{"lineqAGP"} models
}
\references{
F. Bachoc, A. F. Lopez-Lopera, and O. Roustant (2020),
"Sequential construction and dimension reduction of Gaussian processes under inequality constraints".
\emph{ArXiv e-prints}
<arXiv:2009.04188>

F. Bachoc, A. F. Lopez-Lopera, and O. Roustant (2020),
"Sequential construction and dimension reduction of Gaussian processes under inequality constraints".
\emph{ArXiv e-prints}
<arXiv:2009.04188>
}
\author{
A. F. Lopez-Lopera
}
