% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lineqBAGPutils.R
\name{inclusion}
\alias{inclusion}
\title{Indicate if set1 is in set2}
\usage{
inclusion(set1, set2)
}
\arguments{
\item{set1, set2}{two vectors representing sets}
}
\value{
TRUE if set1 is included in set2  FALSE otherwise
}
\description{
Give a boolean saying if set1 is in set2
}
\author{
M. Deronzier and A. F. Lopez-Lopera
}
