% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lineqBAGPutils.R
\name{inv_lemma}
\alias{inv_lemma}
\title{Matrix Inversion Lemma}
\usage{
inv_lemma(U, V, W, invZ)
}
\arguments{
\item{U}{An \eqn{n \times m} matrix}

\item{V}{An \eqn{n \times m} matrix}

\item{W}{An \eqn{m \times m} inversible matrix}

\item{invZ}{An \eqn{n \times n} inversible matrix}
}
\value{
\eqn{(Z + U W V^\top)^{-1}}
}
\description{
Compute \eqn{(Z + U W V^\top)^{-1}} via the matrix Inversion Lemma
(Rasmussen and Williams, 2015; Appendix A.3).
}
\references{
C. E. Rasmussen and C. K. I. Williams (2005),
"Gaussian Processes for Machine Learning".
\emph{The MIT Press}
}
\author{
M. Deronzier and A. F. Lopez-Lopera
}
