% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernels.R, R/kernels_black.R
\name{k1exponential}
\alias{k1exponential}
\title{1D Exponential Kernel Matrix for \code{'lineqGP'} Models.}
\usage{
k1exponential(x1, x2, par, d = 1)

k1exponential(x1, x2, par, d = 1)
}
\arguments{
\item{x1}{a vector with the first input locations.}

\item{x2}{a vector with the second input locations.}

\item{par}{the values of the kernel parameters (variance, lengthscale).}

\item{d}{a number corresponding to the dimension of the input space.}
}
\value{
Kernel matrix \eqn{K(x_1,x_2)}{K(x1,x2)}
(or \eqn{K(x_1,x_1)}{K(x1,x1)} if \eqn{x_2}{x2} is not defined).

Kernel matrix \eqn{K(x_1,x_2)}{K(x1,x2)}
(or \eqn{K(x_1,x_1)}{K(x1,x1)} if \eqn{x_2}{x2} is not defined).
}
\description{
Compute the 1D Exponential kernel for \code{'lineqGP'} models.
attr: 'gradient'.

Compute the 1D Exponential kernel for \code{'lineqGP'} models.
attr: 'gradient'.
}
\examples{
x <- seq(0, 1, 0.01)
K <- k1exponential(x, x, par =  c(1, 0.1))
image(K, main = 'covariance matrix using a Exponential kernel')

x <- seq(0, 1, 0.01)
K <- k1exponential(x, x, par =  c(1, 0.1))
image(K, main = 'covariance matrix using a Exponential kernel')

}
\author{
A. F. Lopez-Lopera
}
