% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lineqGP.R
\name{lineqGPSys}
\alias{lineqGPSys}
\title{Linear Systems of Inequalities for \code{"lineqGP"} Models}
\usage{
lineqGPSys(
  m = nrow(A),
  constrType = c("boundedness", "monotonicity", "decreasing", "convexity", "linear",
    "none"),
  l = -Inf,
  u = Inf,
  A = diag(m),
  d = length(m),
  lineqSysType = c("twosides", "oneside"),
  constrIdx = seq(length(m)),
  rmInf = TRUE
)
}
\arguments{
\item{m}{the number of linear inequality constraints}

\item{constrType}{a character string corresponding to the type of the inequality constraint
Options: "boundedness", "monotonicity", "convexity", "linear"}

\item{l}{the value (or vector) with the lower bound}

\item{u}{the value (or vector) with the upper bound}

\item{A}{a matrix containing the structure of the linear equations}

\item{d}{the value with the input dimension}

\item{lineqSysType}{a character string corresponding to the type of the
linear system. Options: \code{twosides}, \code{oneside} (see \code{\link{bounds2lineqSys}} for more details)}

\item{constrIdx}{for d > 1, a logical vector with the indices of active constrained dimensions}

\item{rmInf}{If \code{TRUE}, inactive constraints are removed
(e.g. \eqn{-\infty \leq x \leq \infty}{-Inf \le x \le Inf}).}
}
\value{
A list with the linear system of inequalities: \code{list(A,l,u)} (\code{twosides}) or \code{list(M,g)} (\code{oneside}).
}
\description{
Build the linear system of inequalities for \code{"lineqGP"} models.
}
\section{Comments}{

This function could change in future versions for more types of inequality
constraints in higher dimensions.
}

\examples{
linSys1 <- lineqGPSys(m = 5, constrType = "boundedness", l = 0, u = 1, lineqSysType = "twosides")
linSys1
linSys2 <- lineqGPSys(m = 5, constrType = "boundedness", l = 0, u = 1, lineqSysType = "oneside")
linSys2

}
\seealso{
\code{\link{bounds2lineqSys}}
}
\author{
A. F. Lopez-Lopera
}
