% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lineqGPutils.R
\name{listBlocks}
\alias{listBlocks}
\title{List of blocks}
\usage{
listBlocks(partition)
}
\arguments{
\item{partition}{a list containing a partition of set $\{1,cdots,D\}$}
}
\value{
a list structure according to the partition.
}
\description{
Create a list according to a given a partition of set $\{1,cdots,D\}$.
}
\examples{
partition <- list(c(1,3), 2)
listBlocks(partition)
}
\author{
A. F. Lopez-Lopera
}
