% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lineqGPlikelihoods.R
\name{logLikBlockAdditiveGrad}
\alias{logLikBlockAdditiveGrad}
\title{Gradient of the Log-Likelihood of a Block Additive Gaussian Process.}
\usage{
logLikBlockAdditiveGrad(
  par = unlist(purrr::map(model$kernParam, "par")),
  model,
  parfixed = rep(FALSE, model$d * length(par)),
  mcmc.opts = NULL,
  estim.varnoise = FALSE
)
}
\arguments{
\item{par}{the values of the covariance parameters.}

\item{model}{an object with \code{"lineqBAGP"} S3 class.}

\item{parfixed}{indices of fixed parameters to do not be optimised.}

\item{mcmc.opts}{not used.}

\item{estim.varnoise}{If \code{true}, a noise variance is estimated.}
}
\value{
the gradient of the negative log-likelihood.
}
\description{
Compute the gradient of the negative log-likelihood of a Block Additive Gaussian Process.
}
\seealso{
\code{\link{logLikBlockAdditiveFun}}
}
\author{
A. F. Lopez-Lopera
}
