% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lineqBAGPutils.R
\name{Mat_E}
\alias{Mat_E}
\title{integral in the general case}
\usage{
Mat_E(subdivision, J = length(subdivision))
}
\arguments{
\item{subdivision}{of type list(list(sequence)) the subdivisions of [0,1]}
}
\value{
the integral fo each hat function described by the subdivision
}
\description{
Give the vector E defined in the article.
}
\author{
M. Deronzier and A. F. Lopez-Lopera
}
