% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lineqBAGPutils.R
\name{matrix_to_block}
\alias{matrix_to_block}
\title{Block operation wrapper function}
\usage{
matrix_to_block(M, subdivision, type = c("bdiag", "cbind", "rbind"))
}
\arguments{
\item{M}{a matrix}

\item{subdivision}{a subdivision of [0,1]}

\item{type}{is a string defining how we want to cut the blocks}
}
\value{
A matrix per block, meaning a list of matrices
}
\description{
The additivity gives us block matrices, hence we can accelerate
the computations this is the role of algorithm in this section to facilitate
the manipulation of matrices of the type \code{list["vector"]}
}
\author{
M. Deronzier and A. F. Lopez-Lopera
}
