% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lineqMaxModBAGP.R
\name{MaxModCriterionBAGP}
\alias{MaxModCriterionBAGP}
\title{MaxMod criterion for \code{"lineqBAGP"} Models}
\usage{
MaxModCriterionBAGP(
  model,
  iter,
  pred,
  option_name,
  option,
  constrType = "none",
  reward_new_dim,
  reward_new_knot,
  activeVar
)
}
\arguments{
\item{model}{an object with class \code{lineqBAGP}}

\item{iter}{an integer corresponding to the iteration of the MaxMod algorithm}

\item{pred}{an object with class \code{lineqBAGP} containing the predictive model}

\item{reward_new_dim}{a number corresponding to the reward of adding a new dimension}

\item{reward_new_knot}{a number corresponding to the reward of adding a new knot in an existing dimension}

\item{subdivision}{a list with the location of the knots}

\item{xtest}{test data for assessing the modification of the MAP estimate}

\item{activeDim}{a sequence containing the active dimensions}

\item{activeDim_IdxSeq}{a sequence containing the indices of the actived input dimensions}

\item{options_expand}{Choice of the algorithm MaxMod}
}
\value{
the value of the knot that minimize the MaxMod criterion and the 
value of the objective function
}
\description{
MaxMod criterion used to add a new knot or a new active dimension (see Bachoc et al., 2020)
}
\details{
#' @title MaxMod criterion for \code{"lineqGP"} Models
#' @description MaxMod criterion used to add a new knot or a new active dimension (see Bachoc et al., 2020)
#' 
#' @param model an object with class \code{lineqGP}
#' @param ulist a list with the location of the knots
#' @param xtest test data for assessing the modification of the MAP estimate
#' @param activeDim a sequence containing the active dimensions
#' @param activeDim_IdxSeq a sequence containing the indices of the actived input dimensions 
#' @param idx_add index of the dimension to be activated
#' @param reward_new_dim a number corresponding to the reward of adding a new dimension
#' @param reward_new_knot a number corresponding to the reward of adding a new knot in an existing dimension
#' @param iter an integer corresponding to the iteration of the MaxMod algorithm 
#' @param pred an object with class \code{lineqGP} containing the predictive model
#' 
#' @return the value of the knot that minimize the MaxMod criterion and the 
#' value of the objective function 
#'
#' @author A. F. Lopez-Lopera
#'
#' @references F. Bachoc, A. F. Lopez-Lopera, and O. Roustant (2020),
#' "Sequential construction and dimension reduction of Gaussian processes under inequality constraints".
#' \emph{ArXiv e-prints}
#' <arXiv:2009.04188>
#'
#' @export
#' 
MaxModCriterion <- function(model, iter, pred , option_name , option,
                            reward_new_dim, reward_new_knot
) {
  class <- class(model)
  fun <- paste("MaxModCriterion.", class, sep = "")
  fun <- try(get(fun))
  if (class(fun) == "try-error") {
    stop('class "', class, '" is not supported')
  } else {
    model <- fun(model, iter, pred , option_name , option,
                 reward_new_dim, reward_new_knot
    )
  }
  return(model)
}
}
\references{
F. Bachoc, A. F. Lopez-Lopera, and O. Roustant (2022),
"Sequential construction and dimension reduction of block additive Gaussian processes under inequality constraints".
}
\author{
A. F. Lopez-Lopera
}
