% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilityFunctions.R
\name{Phi_per_var}
\alias{Phi_per_var}
\title{We can extract evaluation of our GP with a product of matrices $\Phi and \xi$, however, how to 
arrange our function the tensorisation will give us a natural way to arrange things
This function gives the matrice $\Phi$ as a list of matrices per blocks 
Take the vector subdivision and $x$ and get out the object 
$(phi_{{j,k}})(x_{j})_{j,k in}$}
\usage{
Phi_per_var(subdivision, sub_partition, dim_block, X)
}
\value{
Phi_per_var: such that Phi_per_var[[j]][[k]][i] give a seq
of the hat function $\phi_{j,k,i}(X_i)$ where $(X_i)=(x^{(1)}_i,\cdots x^{(n)_i})$ and 
\phi_{j,k,i} is the hat function given by the subdivision[[j]][[k]]

Phi_var_to_tensor: such that Phi_var_to_tensor[[j]] is the 
tensorisation of vectors Phi_per_var[[j]][[k]] for k in $\{1,...,dim_block[j]\}$

Big_Phi is the vector per block defined in the article
}
\description{
We can extract evaluation of our GP with a product of matrices $\Phi and \xi$, however, how to 
arrange our function the tensorisation will give us a natural way to arrange things
This function gives the matrice $\Phi$ as a list of matrices per blocks 
Take the vector subdivision and $x$ and get out the object 
$(phi_{{j,k}})(x_{j})_{j,k in}$
}
