% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lineqBAGPutils.R
\name{Phi_var_to_tensor}
\alias{Phi_var_to_tensor}
\title{Hat basis functions per block}
\usage{
Phi_var_to_tensor(Phi.perVar)
}
\arguments{
\item{Phi.perVar}{A list containing the hat basis functions for each variable.
The matrices are indexed per blocks.}
}
\value{
A list containing the hat basis functions for each block.
}
\description{
Compute the hat basis functions for each block.
}
\author{
M. Deronzier and A. F. Lopez-Lopera
}
