% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lineqGPutils.R
\name{plot.lineqGP}
\alias{plot.lineqGP}
\title{Plot for the \code{"lineqGP"} S3 Class}
\usage{
\method{plot}{lineqGP}(
  x,
  y,
  ytest = NULL,
  probs = c(0.025, 0.975),
  bounds = NULL,
  addlines = TRUE,
  nblines = 5,
  ...
)
}
\arguments{
\item{x}{an object with \code{"lineqGP"} S3 class.}

\item{y}{not used.}

\item{ytest}{the values of the test observations. If \code{!is.null(ytest)}, ytest is drawn.}

\item{probs}{the values of the confidence intervals evaluated at probs.}

\item{bounds}{the values of the bounds of a constrained model. If \code{!is.null(bounds)}, bounds are drawn.}

\item{addlines}{pptional Logical. If \code{TRUE}, some samples are drawn.}

\item{nblines}{if \code{addlines}. The number of samples to be drawn.}

\item{...}{further arguments passed to or from other methods.}
}
\value{
Plot with the \code{"lineqGP"} model.
}
\description{
Plot for the \code{"lineqGP"} S3 class.
}
\seealso{
\code{\link{plot}}, \code{\link{ggplot.lineqGP}}
}
\author{
A. F. Lopez-Lopera
}
