% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lineqBAGPutils.R
\name{seq_int}
\alias{seq_int}
\title{integral in the 1 dimension case}
\usage{
seq_int(subdivision)
}
\arguments{
\item{subdivision}{of type sequence, representing a subdivision of [0,1]}
}
\value{
the integral fo each hat function described by the subdivision
}
\description{
gives the value of integral of  the hat basis functions constructed 
from a 1 dimensional subdivision.
}
\author{
M. Deronzier and A. F. Lopez-Lopera
}
