% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lineqBAGPutils.R
\name{square_norm_int}
\alias{square_norm_int}
\title{Gram matrix of the basis functions (\code{"lineqBAGP"})}
\usage{
square_norm_int(model1, model2)
}
\arguments{
\item{subdivision}{a sequence corresponding to the values of the knots}
}
\value{
Gram matrix of the basis functions
}
\description{
Compute the Gram matrix of the basis functions for \code{"lineqBAGP"} models
}
\references{
F. Bachoc, A. F. Lopez-Lopera, and O. Roustant (2020),
"Sequential construction and dimension reduction of Gaussian processes under inequality constraints".
\emph{ArXiv e-prints}
<arXiv:2009.04188>
}
\author{
M. Deronzier and A. F. Lopez-Lopera
}
