% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lineqBAGPutils.R
\name{sub_Gram}
\alias{sub_Gram}
\title{Gram matrix for the 1 dimensional basis functions (\code{"lineqBAGP"})}
\usage{
sub_Gram(subdivision, n = length(subdivision))
}
\arguments{
\item{subdivision}{a subdivision of [0,1]}

\item{n}{the size of the subdivision}
}
\value{
Gram matrix of the basis functions
}
\description{
Compute the Gram matrix of the basis functions for \code{"lineqBAGP"} models
}
\references{
F. Bachoc, A. F. Lopez-Lopera, and O. Roustant (2020),
"Sequential construction and dimension reduction of Gaussian processes under inequality constraints".
\emph{ArXiv e-prints}
<arXiv:2009.04188>
}
\author{
M. Deronzier and A. F. Lopez-Lopera
}
