% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lineqGPutils.R
\name{tmvrnorm}
\alias{tmvrnorm}
\title{Sampling Methods of Truncated Multivariate Normal Distributions}
\usage{
tmvrnorm(object, nsim, ...)
}
\arguments{
\item{object}{an object with:
\code{mu} (mean vector), \code{Sigma} (covariance matrix),
\code{lb} (lower bound vector), \code{ub} (upper bound vector).}

\item{nsim}{an integer corresponding to the number of simulations.}

\item{...}{further arguments passed to or from other methods.}
}
\value{
A matrix with the sample path. Samples are indexed by columns.
}
\description{
Wrapper function with a collection of Monte Carlo and Markov Chain
Monte Carlo samplers for truncated multivariate normal distributions.
The function invokes particular samplers which depend on the class of the first argument.
}
\seealso{
\code{\link{tmvrnorm.RSM}}, \code{\link{tmvrnorm.HMC}}, 
         \code{\link{tmvrnorm.ExpT}}
}
\author{
A. F. Lopez-Lopera
}
