% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lineqBAGPutils.R
\name{bijection}
\alias{bijection}
\title{Bijection operator}
\usage{
bijection(partition, i)
}
\arguments{
\item{partition}{A list containing the partion of set \eqn{\{1, \cdots, D\}}}

\item{i}{The targeted global input variable}
}
\value{
a sequence containing the indices of the block and the input variable.
}
\description{
A wrapper function used in \code{"lineqBAGP"} models to obtain the indices
of the corresponding block and the local input variable of a global variable. 
The partition \eqn{\{\mathcal{J}_1, \cdots, \mathcal{J}_J \}} of
\eqn{\{1, \cdots, D \}} induces a natural bijection from
\eqn{\{ (j,k), j \in \{1, \cdots, J \}, k \in \{1,\cdots, \mathcal{J}_b\} \} }
into \eqn{\{1, \cdots, D\}} there is an straight one, \code{partition[[j]][[k]]}
gives the element \eqn{i} and ... (anfe: to be complete!!)
}
\examples{
partition <- list(c(1,3), c(2,4,5))
bijection(partition, 3)
}
\author{
M. Deronzier and A. F. Lopez-Lopera
}
