% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lineqBAGPutils.R
\name{block_prod}
\alias{block_prod}
\title{Elementwise product of two lists of matrices}
\usage{
block_prod(A, B)
}
\arguments{
\item{A}{A list containing matrices.}

\item{B}{A list containing matrices.}
}
\value{
A list containing the elementwise product.
}
\description{
Compute the elementwise product of two lists of matrices.
}
\author{
M. Deronzier and A. F. Lopez-Lopera
}
