% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilityFunctions.R
\name{inv_lemma}
\alias{inv_lemma}
\title{We will introduce here the functions that will be able to constructe and manipulate the kernel matrix
The kernel is obtained by tensorisation
We will create functions for different matrices we will neeed
Computation of the conditional covariance matrice using the inversion lemma formulation:
$(Z+UWV^{\top})^{-1}=Z^{-1}-Z^{-1}U(W^{-1}+ V^{t}Z^{-1}U)^{-1}V^{\top}Z^{-1}$ it is useful
when $U, V\in M_{n,m}, W \in M_{m,m} $ and $Z \in M_{n,n}$. the factorisation form is useful
when $m<<n$}
\usage{
inv_lemma(U, V, W, inv_Z)
}
\description{
We will introduce here the functions that will be able to constructe and manipulate the kernel matrix
The kernel is obtained by tensorisation
We will create functions for different matrices we will neeed
Computation of the conditional covariance matrice using the inversion lemma formulation:
$(Z+UWV^{\top})^{-1}=Z^{-1}-Z^{-1}U(W^{-1}+ V^{t}Z^{-1}U)^{-1}V^{\top}Z^{-1}$ it is useful
when $U, V\in M_{n,m}, W \in M_{m,m} $ and $Z \in M_{n,n}$. the factorisation form is useful
when $m<<n$
}
