% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilityFunctions.R
\name{Phi_var_to_tensor}
\alias{Phi_var_to_tensor}
\title{This function take the phi per variables and get back the family $\phi_{\el_j}(x_{j})$}
\usage{
Phi_var_to_tensor(Phi_per_var, subdivision, subdivision_size, dim_block)
}
\description{
This function take the phi per variables and get back the family $\phi_{\el_j}(x_{j})$
}
